+++ RAnalogs/RAnEn/R/verify.R
@@ -159,6 +159,6 @@ verify <- function(metrics, verbose = T, ...) {
   
   rm(args.all, args.required, args.optional, args.current)
   garbage <- gc(reset = T)
-  cat('Done\n')
+  cat('Done (verify)\n')
   return(ret)
 }